+++ vaultpkitest/vaultpkitest.go
@@ -1,6 +1,7 @@
 package vaultpkitest
 
 import (
+	"github.com/giantswarm/vaultpki"
 	vaultapi "github.com/hashicorp/vault/api"
 )
 
@@ -23,8 +24,12 @@ func (p *VaultPKITest) CreateBackend(ID string) error {
 	return nil
 }
 
-func (p *VaultPKITest) CreateCA(ID string) error {
-	return nil
+func (p *VaultPKITest) CreateCA(ID string) (vaultpki.CertificateAuthority, error) {
+	return vaultpki.DefaultCertificateAuthority(), nil
+}
+
+func (p *VaultPKITest) CreateCAWithPrivateKey(ID string) (vaultpki.CertificateAuthority, error) {
+	return vaultpki.DefaultCertificateAuthority(), nil
 }
 
 func (p *VaultPKITest) DeleteBackend(ID string) error {
@@ -35,8 +40,8 @@ func (p *VaultPKITest) GetBackend(ID string) (*vaultapi.MountOutput, error) {
 	return nil, nil
 }
 
-func (p *VaultPKITest) GetCACertificate(ID string) (string, error) {
-	return "", nil
+func (p *VaultPKITest) GetCACertificate(ID string) (vaultpki.CertificateAuthority, error) {
+	return vaultpki.DefaultCertificateAuthority(), nil
 }
 
 func (p *VaultPKITest) ListBackends() (map[string]*vaultapi.MountOutput, error) {